refactor/fix: map backend status codes to user-friendly messages; move auth pages to their own folder
