+++ frontend/src/pages/index.ts
@@ -3,8 +3,8 @@ export { default as Home } from './Home';
 export { default as NotFound } from './NotFound';
 export { default as Loading } from './Loading';
 export { default as Landing } from './Landing';
-export { default as Registration } from './Registration';
-export { default as Login } from './Login';
+export { default as Registration } from './auth/Registration';
+export { default as Login } from './auth/Login';
 export { default as ProtectedRoute } from './ProtectedRoute';
 export { default as Chat } from './Chat';
 export { default as TopRightAvatar } from './TopRightAvatar';
